==> step03_spatial_joins.Rmd <==
---
title: "Spatial Joins"
author: "Amanda Gassett"
date: "2024-04-27"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(sf)
```


```{r, echo = FALSE}
g1 <- read.csv("../cleannicu96_geocoded.csv")
g2 <- read.csv("../cleannicu11_geocoded.csv")
g3 <- read.csv("../cleannicu23_geocoded.csv")
all_locs <- unique(rbind.data.frame(g1[,c("hosp_id","lat","lon")],
                             g2[,c("hosp_id","lat","lon")],
                             g3[,c("hosp_id","lat","lon")]))

nicu96.geo <- st_as_sf(read.csv("../cleannicu96_geocoded.csv"), coords = c("lon", "lat"), crs = 4326)
nicu11.geo <- st_as_sf(read.csv("../cleannicu11_geocoded.csv"), coords = c("lon", "lat"), crs = 4326)
nicu23.geo <- st_as_sf(read.csv("../cleannicu23_geocoded.csv"), coords = c("lon", "lat"), crs = 4326)

# Find nearest 2011 record to 1996
m1 <- nicu23.geo[st_nearest_feature(nicu11.geo, nicu23.geo),c("name","full_address","beds","level")]
m1$dist_11_to_23 <- as.numeric(st_distance(nicu11.geo, m1, by_element = TRUE))
names(m1)[1:4] <- c("name23","address23","beds23","level23")

wide11 <- cbind.data.frame(nicu11.geo[,c("hosp_name","full_address","beds","level")],
                           m1)

# Find nearest 2011 record to 1996
m1 <- nicu11.geo[st_nearest_feature(nicu96.geo, nicu11.geo),c("hosp_name","address","level","beds","num_drs")]
m1$dist_96_to_11 <- as.numeric(st_distance(nicu96.geo, m1, by_element = TRUE))
names(m1)[1:5] <- c("name11","address11","level11","beds11","num_drs11")

wide96 <- cbind.data.frame(nicu96.geo[,c("hosp_name","address","city","state","beds","num_drs")],
                           m1)

# Find nearest 2023 record to 1996
m2 <- nicu23.geo[st_nearest_feature(nicu96.geo, nicu23.geo),c("name","level","beds","address","type")]
m2$dist_96_to_23 <- as.numeric(st_distance(nicu96.geo, m2, by_element = TRUE))
names(m2)[1:5] <- c("name23","level23","beds23","address23","type23")
wide96 <- cbind.data.frame(wide96, m2)

dim(wide96[as.numeric(wide96$dist_96_to_11) > 25000 | as.numeric(wide96$dist_96_to_23) > 25000,])


write.csv(wide96[as.numeric(wide96$dist_96_to_11) > 25000 | as.numeric(wide96$dist_96_to_23) > 25000,
                 c("hosp_name","name11","name23",
                   "dist_96_to_11","dist_96_to_23",
                   "address","address11","address23",
                   "beds","beds11","beds23")], "../moves96.csv", row.names = FALSE)

write.csv(wide96[as.numeric(wide96$dist_96_to_11) < 500 & as.numeric(wide96$dist_96_to_23) < 500 & 
                   (is.na(wide96$beds) | is.na(wide96$beds11) | is.na(wide96$beds23)),
                 c("hosp_name","name11","name23",
                   "dist_96_to_11","dist_96_to_23",
                   "address","address11","address23",
                   "level11","level23","num_drs","num_drs11",
                   "beds","beds11","beds23")], "../na_bed_counts.csv", row.names = FALSE)

# Find nearest 2011 record to 2023
m2 <- nicu11.geo[st_nearest_feature(nicu23.geo, nicu11.geo),c("hosp_name","level","beds","address")]
m2$dist_23_to_11 <- as.numeric(st_distance(nicu23.geo, m2, by_element = TRUE))
names(m2)[1:4] <- c("name11","level11","beds11","address11")
wide23 <- cbind.data.frame(nicu23.geo, m2)
table(wide23[,c("level11","level")])

# CMS-Based Neonatal units
cms_based <- read.csv("../extra_line_units.csv")
cms.geo <- st_as_sf(cms_based[cms_based$unit_type == "NEO", c("hospital_name","address","city","state","zip_code","hospital_type","lat","lng","geocode_flag","geocode_accuracy","total")],
                    coords = c("lng", "lat"), crs = 4326)
m3 <- cms.geo[st_nearest_feature(nicu23.geo, cms.geo),]
m3$cms_dist <- as.numeric(st_distance(nicu23.geo, m3, by_element = TRUE))
wider23 <- cbind.data.frame(wide23, m3)



```
```{r}
# CMS-Based total beds
cms_based2 <- read.csv("../hospital_data.csv")
cms_based2 <- cms_based2[!is.na(cms_based2$lat),]
cms.geo2 <- st_as_sf(cms_based2[, c("hospital_name","address","city","state","zip_code","hospital_type","lat","lng",
                                                               "geocode_flag","geocode_accuracy","labor_delivery_beds_3200","all_beds_2700")],
                    coords = c("lng", "lat"), crs = 4326)
m4 <- cms.geo2[st_nearest_feature(nicu23.geo, cms.geo2),]
m4$cms_dist2 <- as.numeric(st_distance(nicu23.geo, m4, by_element = TRUE))
widest23 <- cbind.data.frame(wider23, m4)

plot(total~beds, data = widest23[widest23$cms_dist < 500,], col = as.factor(level),
     xlab = "Neonatology Solutions", ylab = "CMS-Based Neonatal Bed Count")
legend("bottomright",col = 1:3, pch = c(1,1,1), legend = levels(as.factor(widest23$level)))
abline(0,1)

plot(labor_delivery_beds_3200~beds, data = widest23[widest23$cms_dist2 < 500,], col = as.factor(level),
     xlab = "Neonatology Solutions", ylab = "CMS-Based Labor/Delivery")
legend("bottomright",col = 1:3, pch = c(1,1,1), legend = levels(as.factor(widest23$level)))

plot(all_beds_2700~beds, data = widest23[widest23$cms_dist2 < 500,], col = as.factor(level),
     xlab = "Neonatology Solutions", ylab = "CMS-Based All Beds")
legend("bottomright",col = 1:3, pch = c(1,1,1), legend = levels(as.factor(widest23$level)))


plot(labor_delivery_beds_3200~all_beds_2700, data = widest23[widest23$cms_dist2 < 500,])
plot(beds~all_beds_2700, data = widest23[widest23$cms_dist2 < 500,])
plot(beds~beds11, data = widest23[widest23$dist_23_to_11 < 500,])

summary(lm(beds ~ as.factor(level) + all_beds_2700 + labor_delivery_beds_3200, data = widest23[widest23$cms_dist2 < 500,]))
summary(lm(beds ~ as.factor(level)*all_beds_2700, data = widest23[widest23$cms_dist2 < 500,]))
temp2 <- predict(lm(beds ~ as.factor(level)*all_beds_2700, data = widest23[widest23$cms_dist2 < 500,]), widest23)
plot(temp2 ~ widest23$beds)


summary(lm(beds ~ as.factor(level), data = widest23))
summary(lm(beds ~ as.factor(level)*as.factor(state), data = widest23))
temp <- predict(lm(beds ~ as.factor(level)*as.factor(state), data = widest23), widest23)
plot(temp ~ widest23$beds)

write.csv(widest23[,!grepl("geom", names(widest23))], "../nicu23_impute.csv", row.names = FALSE, na = "")

```

The table below shows level designations for 2011 and 2023, for facilities that existed in the same location for all 3 years (i.e. geocode within 500m). Note that the table below does not account for some duplicates where there might be e.g. a level 2 and level 4 NICU in the same facility but provided as separate directory records.

```{r}
table(wide96[as.numeric(wide96$dist_96_to_11) < 500 & as.numeric(wide96$dist_96_to_23) < 500,c("level11","level23")], useNA = "always")

plot(beds ~ beds11, ylab = "Bed Count 1996", xlab = "Bed Count 2011",
     data = wide96[as.numeric(wide96$dist_96_to_11) < 500 & as.numeric(wide96$dist_96_to_23) < 500,],
     bg = is.na(beds23) + 2, pch = 23)
legend("bottomright",legend = c("Not Missing","Bed Count '23 Missing"), pt.bg = c(2,3), pch = c(23,23))

plot(beds ~ beds23, ylab = "Bed Count 1996", xlab = "Bed Count 2023",
     data = wide96[as.numeric(wide96$dist_96_to_11) < 500 & as.numeric(wide96$dist_96_to_23) < 500,],
     bg = is.na(beds11) + 2, pch = 23)
legend("bottomright",legend = c("Not Missing","Bed Count '11 Missing"), pt.bg = c(2,3), pch = c(23,23))

plot(beds11 ~ beds23, ylab = "Bed Count 2011", xlab = "Bed Count 2023",
     data = wide96[as.numeric(wide96$dist_96_to_11) < 500 & as.numeric(wide96$dist_96_to_23) < 500,],
     bg = is.na(beds) + 2, pch = 23)
legend("bottomright",legend = c("Not Missing","Bed Count '96 Missing"), pt.bg = c(2,3), pch = c(23,23))


```
